pyproject.toml
[tool.poetry]
name = "org-structure"
version = "0.1.0"
description = ""
authors = ["DANYA <daniiildolgov@yandex.ru>"]

[tool.poetry.dependencies]
python = "^3.11"
alembic = "1.18.4"
annotated-doc = "0.0.4"
annotated-types = "0.7.0"
anyio = "4.12.1"
fastapi = "0.129.0"
greenlet = "3.3.1"
idna = "3.11"
mako = "1.3.10"
markupsafe = "3.0.3"
pydantic = "2.12.5"
pydantic-settings = "2.13.1"
pydantic-core = "2.41.5"
python-dotenv = "1.2.1"
sqlalchemy = "2.0.46"
starlette = "0.52.1"
typing-inspection = "0.4.2"
typing-extensions = "4.15.0"
uvicorn = ">=0.41.0,<0.42.0"
asyncpg = ">=0.31.0,<0.32.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
[dependency-groups]
dev = [
    "pytest (>=9.0.2,<10.0.0)",
    "pytest-asyncio (>=1.3.0,<2.0.0)",
    "httpx (>=0.28.1,<0.29.0)",
    "pytest-env (>=1.5.0,<2.0.0)"
]
